Fix 'README.md' examples for the  'ctx' parameter
The 'ctx' parameter accepts one of two values: 'thread' to run the scheduled job on a separate thread, which is the default behaviour if no 'ctx' value is specified, or 'multiprocess' to run the scheduled job on a separate process. The 'README.md' file incorrectly showed 'process' as a valid value for the 'ctx' parameter and none of the correct values mentioned before.

Fixes: #254