Remove old py26 compatibility logging
This code was creating loggers in py26+ versions that had the name
'root.crython...' instead of just 'crython.'. Since py26 is no longer
supported, it should be safe to remove this code. It's a backwards
compatbility breaking change however, for anyone that had figured
that out and used 'root.' in their logging configs.